+++ src/main/java/com/codecool/crudinventorymanagementapp/employee/EmployeeModel.java
@@ -3,13 +3,14 @@ package com.codecool.crudinventorymanagementapp.employee;
 import org.hibernate.validator.constraints.NotEmpty;
 
 import javax.persistence.*;
+import java.math.BigInteger;
 
 @Entity
 public class EmployeeModel {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @NotEmpty
     private String name;
@@ -36,11 +37,11 @@ public class EmployeeModel {
 
     @NotEmpty
     private String login;
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
